chore: update SwiftTestMockProvider implementation to MockProvider
