+++ Tests/RudderStackAnalyticsTests/EventsModule/EventProcessing/EventQueueTests.swift
@@ -18,7 +18,7 @@ class EventQueueTests {
     
     init() {
         mockStorage = MockStorage()
-        let config = SwiftTestMockProvider.createMockConfiguration(storage: mockStorage)
+        let config = MockProvider.createMockConfiguration(storage: mockStorage)
         config.trackApplicationLifecycleEvents = false
         config.sessionConfiguration.automaticSessionTracking = false
         config.flushPolicies = []
