+++ Tests/RudderStackAnalyticsTests/Network/HttpClientTests.swift
@@ -16,7 +16,7 @@ struct HttpClientTests {
     private let httpClient: HttpClient
 
     init() {
-        mockAnalytics = SwiftTestMockProvider.createMockAnalytics()
+        mockAnalytics = MockProvider.createMockAnalytics()
         httpClient = HttpClient(analytics: mockAnalytics)
     }
 
@@ -91,8 +91,8 @@ struct HttpClientTests {
 
     @Test("given successful HTTP response, when requesting configuration data, then configuration data is returned successfully")
     func testGetConfigDataSuccess() async {
-        SwiftTestMockProvider.setupMockURLSession()
-        defer { SwiftTestMockProvider.teardownMockURLSession() }
+        MockProvider.setupMockURLSession()
+        defer { MockProvider.teardownMockURLSession() }
         
         let expectedData = Data("{\"success\": true}".utf8)
         MockURLProtocol.requestHandler = { request in
@@ -105,8 +105,8 @@ struct HttpClientTests {
     
     @Test("given a failure HTTP response, when requesting configuration data, then the error is handled properly")
     func testGetConfigDataFailure() async {
-        SwiftTestMockProvider.setupMockURLSession()
-        defer { SwiftTestMockProvider.teardownMockURLSession() }
+        MockProvider.setupMockURLSession()
+        defer { MockProvider.teardownMockURLSession() }
         
         MockURLProtocol.requestHandler = { request in
             return (400, nil, nil)
@@ -118,8 +118,8 @@ struct HttpClientTests {
     
     @Test("given successful HTTP response, when posting batch events, then handles success response")
     func testPostBatchEventsSuccess() async {
-        SwiftTestMockProvider.setupMockURLSession()
-        defer { SwiftTestMockProvider.teardownMockURLSession() }
+        MockProvider.setupMockURLSession()
+        defer { MockProvider.teardownMockURLSession() }
         
         let eventBatch = "{\"batch\": [\"event1\", \"event2\"]}"
         let expectedResponseData = "{\"success\": true}".utf8Data
@@ -134,8 +134,8 @@ struct HttpClientTests {
     
     @Test("given a failure HTTP response, when posting batch events failure, then the error is handled properly")
     func testPostBatchEventsFailure() async {
-        SwiftTestMockProvider.setupMockURLSession()
-        defer { SwiftTestMockProvider.teardownMockURLSession() }
+        MockProvider.setupMockURLSession()
+        defer { MockProvider.teardownMockURLSession() }
         
         let batchData = "{\"batch\": []}"
         MockURLProtocol.requestHandler = { request in
