+++ Tests/RudderStackAnalyticsTests/Plugins/Basic/LibraryInfoPluginTests.swift
@@ -17,14 +17,14 @@ class LibraryInfoPluginTests {
     }
     
     @Test("when intercepting different events, then adds library context information", arguments:[
-        SwiftTestMockProvider.mockTrackEvent as Event,
-        SwiftTestMockProvider.mockScreenEvent as Event,
-        SwiftTestMockProvider.mockIdentifyEvent as Event,
-        SwiftTestMockProvider.mockGroupEvent as Event,
-        SwiftTestMockProvider.mockAliasEvent as Event
+        MockProvider.mockTrackEvent as Event,
+        MockProvider.mockScreenEvent as Event,
+        MockProvider.mockIdentifyEvent as Event,
+        MockProvider.mockGroupEvent as Event,
+        MockProvider.mockAliasEvent as Event
     ])
     func test_pluginIntercept(_ event: Event) {
-        let analytics = SwiftTestMockProvider.createMockAnalytics()
+        let analytics = MockProvider.createMockAnalytics()
         libraryInfoPlugin.setup(analytics: analytics)
         
         let result = libraryInfoPlugin.intercept(event: event)
@@ -48,7 +48,7 @@ class LibraryInfoPluginTests {
     
     @Test("when setup is called, then analytics reference is stored")
     func test_pluginSetup() {
-        let analytics = SwiftTestMockProvider.createMockAnalytics()
+        let analytics = MockProvider.createMockAnalytics()
         
         libraryInfoPlugin.setup(analytics: analytics)
