+++ Tests/RudderStackAnalyticsTests/Plugins/Basic/TimeZoneInfoPluginTests.swift
@@ -17,14 +17,14 @@ class TimeZoneInfoPluginTests {
     }
     
     @Test("when intercepting different events, then adds timezone context information", arguments:[
-        SwiftTestMockProvider.mockTrackEvent as Event,
-        SwiftTestMockProvider.mockScreenEvent as Event,
-        SwiftTestMockProvider.mockIdentifyEvent as Event,
-        SwiftTestMockProvider.mockGroupEvent as Event,
-        SwiftTestMockProvider.mockAliasEvent as Event
+        MockProvider.mockTrackEvent as Event,
+        MockProvider.mockScreenEvent as Event,
+        MockProvider.mockIdentifyEvent as Event,
+        MockProvider.mockGroupEvent as Event,
+        MockProvider.mockAliasEvent as Event
     ])
     func test_pluginIntercept(_ event: Event) {
-        let analytics = SwiftTestMockProvider.createMockAnalytics()
+        let analytics = MockProvider.createMockAnalytics()
         timeZoneInfoPlugin.setup(analytics: analytics)
         
         let result = timeZoneInfoPlugin.intercept(event: event)
@@ -41,7 +41,7 @@ class TimeZoneInfoPluginTests {
     
     @Test("when setup is called, then analytics reference is stored")
     func test_pluginSetup() {
-        let analytics = SwiftTestMockProvider.createMockAnalytics()
+        let analytics = MockProvider.createMockAnalytics()
         
         timeZoneInfoPlugin.setup(analytics: analytics)
